+++ Content/Interactables/Scrapper.cs
@@ -91,7 +91,7 @@ namespace WellRoundedBalance.Interactables
         private void ClassicStageInfo_Start(On.RoR2.ClassicStageInfo.orig_Start orig, ClassicStageInfo self)
         {
             orig(self);
-            if (NetworkServer.active)
+            if (NetworkServer.active && self.interactableCategories)
             {
                 var categories = self.interactableCategories.categories;
                 for (int i = 0; i < categories.Length; i++)