+++ Content/Items/Whites/PowerElixir.cs
@@ -17,10 +17,10 @@ namespace WellRoundedBalance.Items.Whites
         public override string DescText => $"Taking damage to below <style=cIsHealth>{d(healthThreshold)} health</style> <style=cIsUtility>consumes</style> this item, <style=cIsHealing>healing</style> you for <style=cIsHealing>{d(percentHealing)}</style> of your <style=cIsHealing>maximum health</style> over <style=cIsUtility>{healingTime}s</style>." +
             (refillEveryStage ? " <style=cIsUtility>Refills every stage</style>." : "");
 
-        [ConfigField("Health Threshold", "Decimal.", 0.5f)]
+        [ConfigField("Health Threshold", "Decimal.", 0.35f)]
         public static float healthThreshold;
 
-        [ConfigField("Percent Healing", "Decimal.", 0.2f)]
+        [ConfigField("Percent Healing", "Decimal.", 0.4f)]
         public static float percentHealing;
 
         [ConfigField("Healing Time", 4f)]